Create ollivander_wand
Problem: Harry Potter and his friends are at Ollivander's with Ron, finally replacing Charlie's old broken wand.
Hermione decides the best way to choose is by determining the minimum number of gold galleons needed to buy each non-evil wand of high power and age. Write a query to print the id, age, coins_needed, and power of the wands that Ron's interested in, sorted in order of descending power. If more than one wand has same power, sort the result in order of descending age. Mapping between code and age is 1:1
Solution: Rank to find in each code and power which IDs need minimum coins and then filter on rank = 1 to get the list and sort as requested